+++ BackEnd/Services/Interfaces/IOrderService.cs
@@ -1,6 +1,4 @@
 ﻿using DAL.InputModels;
-using DAL.ViewModels;
-using System.Collections.Generic;
 
 namespace Services
 {
@@ -9,13 +7,5 @@ namespace Services
         string StartOrder(OrderDetailsInputModel orderDetails, string userId);
 
         void JionOrder(string orderId, string userId);
-
-        string GetActiveOrder(string userId);
-
-        IEnumerable<MenuItemBoardViewModel> GetMenu(string userId);
-
-        void AddItemToOrder(MenuItemOrderInputModel item, string userId);
-
-        IEnumerable<OrderedMenuItemBoardViewModel> GetOrderedItems(string userId);
     }
 }
